+++ src/components/sidebar/Sidebar.jsx
@@ -1,57 +1,30 @@
-import {
-  Avatar,
-  Badge,
-  Button,
-  Icon,
-  IconButton,
-  Image,
-  Link,
-  Text,
-  useMediaQuery,
-} from '@chakra-ui/react'
+import { useMediaQuery } from '@chakra-ui/react'
 import { useEffect, useState } from 'react'
-import { FaPlay } from 'react-icons/fa'
-import {
-  SiDiscord,
-  SiFacebook,
-  SiInstagram,
-  SiTwitter,
-  SiYoutube,
-} from 'react-icons/si'
 import { useDispatch, useSelector } from 'react-redux'
-import { Link as ReactRouterLink } from 'react-router-dom'
 
-import logoFull from '@assets/images/logo_symbol_full.svg'
-import logoSymbol from '@assets/images/logo_symbol_white.svg'
 import {
-  BackpackIcon,
-  BellFilledIcon,
-  BlockIcon,
-  ClockIcon,
   Container,
-  ExitIcon,
   FriendList,
-  FriendsIcon,
-  JoystickIcon,
   LogoutModal,
   NotificationList,
-  PlayIcon,
-  PodiumIcon,
-  SettingsIcon,
-  ShareIcon,
-  ShopIcon,
-  SupportIcon,
+  SidebarAvatarLink,
+  SidebarFooter,
+  SidebarLobbyButton,
+  SidebarLogo,
+  SidebarMenuItem,
   SupportModal,
-  Timer,
-  UserIcon,
 } from '@components'
 import { toggleFriendList } from '@slices/AppSlice'
-import { formatSecondsToMinutes } from '@utils'
 
 import style from './Sidebar.module.css'
 
-export default function Sidebar({ collapsed = true, collapsable = false }) {
+const topMenuItems = ['amigos', 'notificações', 'ranking', 'loja']
+const bottomMenuItems = ['suporte', 'sair']
+
+export default function Sidebar() {
   const [isLessThan2xl] = useMediaQuery('(max-width: 1600px)')
+
+  const dispatch = useDispatch()
 
   const user = useSelector((state) => state.user)
   const lobby = useSelector((state) => state.lobby)
@@ -60,18 +33,18 @@ export default function Sidebar({ collapsed = true, collapsable = false }) {
   const invites = useSelector((state) => state.invites)
   const friendListOpenByApp = useSelector((state) => state.app.friendListOpen)
 
-  const dispatch = useDispatch()
-
-  const [isCollapsed, setIsCollapsed] = useState(collapsable && collapsed)
   const [openSupport, setOpenSupport] = useState(false)
   const [openLogoutModal, setOpenLogoutModal] = useState(false)
   const [notificationsOpen, setNotificationsOpen] = useState(false)
   const [friendListOpen, setFriendListOpen] = useState(false)
-  const [unreadNotificationsCount, setUnreadNotificationsCount] = useState(0)
 
   const receivedInvites = invites.filter(
     (invite) => invite.to_player.user_id === user.id
-  )
+  ).length
+
+  const unreadNotifications = notifications.filter(
+    (notification) => notification.read_date === null
+  ).length
 
   const handleOpenModalSupport = () => {
     setOpenSupport(true)
@@ -79,103 +52,6 @@ export default function Sidebar({ collapsed = true, collapsable = false }) {
 
   const handleOpenModalLogout = () => {
     setOpenLogoutModal(true)
-  }
-
-  const open = () => {
-    collapsable && setIsCollapsed(false)
-  }
-
-  const collapse = () => {
-    collapsable && setIsCollapsed(true)
-  }
-
-  const renderButtons = () => {
-    const showPlayButton =
-      !lobby.queue && !match && !lobby.restriction_countdown
-
-    const showQueueButton =
-      lobby.queue && !match && !lobby.restriction_countdown
-
-    const showRestrictedButton = lobby.restriction_countdown && !match
-
-    return (
-      <>
-        {showPlayButton && (
-          <Button
-            leftIcon={<FaPlay />}
-            className={style.playBtn}
-            fontSize={{ base: 18, md: 16, '2xl': 18 }}
-            fontWeight="bold"
-            height={{ base: 55, md: 50, '2xl': 55 }}
-            w="full"
-            as={ReactRouterLink}
-            to="/jogar"
-          >
-            COMEÇAR
-          </Button>
-        )}
-
-        {showQueueButton && (
-          <Button
-            leftIcon={<ClockIcon />}
-            className={style.queueBtn}
-            fontSize={{ base: 18, md: 16, '2xl': 18 }}
-            fontWeight="bold"
-            height={{ base: 55, md: 50, '2xl': 55 }}
-            w="full"
-            as={ReactRouterLink}
-            to="/jogar"
-            variant="queue"
-          >
-            <Text minW="52px" top="1px" position="relative">
-              {formatSecondsToMinutes(lobby.queue_time)}
-            </Text>
-          </Button>
-        )}
-
-        {match && (
-          <Button
-            leftIcon={<JoystickIcon />}
-            className={style.queueBtn}
-            fontSize={{ base: 18, md: 16, '2xl': 18 }}
-            fontWeight="bold"
-            height={{ base: 55, md: 50, '2xl': 55 }}
-            w="full"
-            as={ReactRouterLink}
-            to={`partidas/${match.id}`}
-            variant="queue"
-          >
-            EM PARTIDA
-          </Button>
-        )}
-
-        {showRestrictedButton && (
-          <Button
-            className={style.dangerBtn}
-            fontSize={{ base: 18, md: 16, '2xl': 18 }}
-            fontWeight="bold"
-            height={{ base: 55, md: 50, '2xl': 55 }}
-            w="full"
-            as={ReactRouterLink}
-            to="/jogar"
-            variant="restricted"
-          >
-            <Container column align="center" gap={4}>
-              <Text fontSize={12} color="white" fontWeight="semiBold">
-                GRUPO COM RESTRIÇÃO
-              </Text>
-
-              <Container justify="center" align="center" gap={4}>
-                <Icon as={BlockIcon} fill="white" w="16px" h="16px" />
-                <Text fontSize={18} fontWeight="bold" w={10} lineHeight={1}>
-                  <Timer initialTime={lobby.restriction_countdown} reverse />
-                </Text>
-              </Container>
-            </Container>
-          </Button>
-        )}
-      </>
-    )
   }
 
   const handleToggleNotificationsDrawer = () => {
@@ -198,100 +74,49 @@ export default function Sidebar({ collapsed = true, collapsable = false }) {
     } else setFriendListOpen(true)
   }
 
-  useEffect(() => {
-    if (collapsable) setIsCollapsed(collapsed)
-    else setIsCollapsed(false)
-  }, [collapsed, collapsable])
+  const onClickFunction = (item) => {
+    switch (item) {
+      case 'amigos':
+        handleToggleFriendListDrawer()
+        break
+
+      case 'notificações':
+        handleToggleNotificationsDrawer()
+        break
+
+      case 'suporte':
+        handleOpenModalSupport()
+        break
+
+      case 'sair':
+        handleOpenModalLogout()
+        break
+
+      default:
+        return null
+    }
+  }
 
   useEffect(() => {
     setFriendListOpen(friendListOpenByApp)
   }, [friendListOpenByApp])
 
-  useEffect(() => {
-    if (notifications && notifications.length > 0) {
-      const notificationsNotRead = notifications.filter(
-        (notification) => notification.read_date === null
-      ).length
-
-      setUnreadNotificationsCount(notificationsNotRead)
-    }
-  }, [notifications])
-
   return (
     <>
       <Container
         column
-        className={[
-          style.container,
-          collapsable && isCollapsed && style.collapsed,
-        ].join(' ')}
+        className={style.container}
         justify="between"
-        onMouseEnter={open}
-        onMouseLeave={collapse}
         testID="container"
-        gap={isLessThan2xl ? 30 : 48}
+        gap={isLessThan2xl ? 58 : 78}
       >
-        <Container column fitContent gap={isLessThan2xl ? 28 : 48}>
-          <Container className={style.logoWrapper}>
-            <Link as={ReactRouterLink} to="/jogar">
-              <Image
-                src={logoSymbol}
-                style={{ height: isCollapsed ? 'auto' : 0 }}
-                data-testid="logo-symbol"
-              />
-              <Image
-                src={logoFull}
-                style={{ height: !isCollapsed ? 'auto' : 0 }}
-                data-testid="logo-full"
-              />
-            </Link>
-          </Container>
-          <Container
-            align="center"
-            justify="center"
-            fitContent
-            className={style.btnWrapper}
-          >
-            {isCollapsed ? (
-              <IconButton
-                as={Button}
-                icon={<PlayIcon />}
-                aria-label="Jogar"
-                fontSize={18}
-                fontWeight="bold"
-                height={55}
-                className={style.playBtn}
-              />
-            ) : (
-              renderButtons()
-            )}
-          </Container>
-        </Container>
+        <SidebarLogo />
 
-        <Container className={style.body} column gap={14}>
-          <Container className={style.userInfo} align="center">
-            <Container gap={14} align="center" justify="center">
-              <Avatar src={user.account.avatar.medium} variant={user.status} />
-              {!isCollapsed && (
-                <Container column>
-                  <Text
-                    color="white"
-                    fontWeight={'medium'}
-                    fontSize={{ base: 16, md: 14, '2xl': 16 }}
-                  >
-                    {user.account.username}
-                  </Text>
+        <Container className={style.body} column gap={isLessThan2xl ? 35 : 50}>
+          <Container className={style.userInfo} gap={24} align="center" column>
+            <SidebarAvatarLink user={user} />
 
-                  <Text
-                    color="gray.700"
-                    fontWeight={'medium'}
-                    fontSize={{ base: 12, md: 10, '2xl': 12 }}
-                  >
-                    LEVEL {user.account.level}
-                  </Text>
-                </Container>
-              )}
-            </Container>
+            <SidebarLobbyButton lobby={lobby} match={match} />
           </Container>
 
           <Container
@@ -300,194 +125,32 @@ export default function Sidebar({ collapsed = true, collapsable = false }) {
             className={style.menu}
           >
             <Container className={style.topMenu} column>
-              <Container className={style.menuItem}>
-                <Link as="button" onClick={handleToggleFriendListDrawer}>
-                  <Container
-                    className={style.menuLinkWrapper}
-                    gap={14}
-                    align="center"
-                  >
-                    <Icon as={FriendsIcon} fill="gray.700" />
-                    {!isCollapsed && <Text>Amigos</Text>}
-                  </Container>
-                  <Container
-                    justify={!isCollapsed ? 'end' : 'start'}
-                    className={style.unreadBadge}
-                  >
-                    <Badge
-                      variant={isCollapsed ? 'unread' : 'counter'}
-                      style={{ opacity: receivedInvites.length > 0 ? 1 : 0 }}
-                    >
-                      {!isCollapsed && receivedInvites.length}
-                    </Badge>
-                  </Container>
-                </Link>
-              </Container>
-
-              <Container className={style.menuItem}>
-                <Link as="button" onClick={handleToggleNotificationsDrawer}>
-                  <Container className={style.menuLinkWrapper} gap={14}>
-                    <Icon as={BellFilledIcon} fill="gray.700" />
-                    {!isCollapsed && <Text fontSize={14}>Notificações</Text>}
-                  </Container>
-
-                  <Container
-                    justify={!isCollapsed ? 'end' : 'start'}
-                    className={style.unreadBadge}
-                  >
-                    <Badge
-                      variant={isCollapsed ? 'unread' : 'counter'}
-                      style={{ opacity: unreadNotificationsCount > 0 ? 1 : 0 }}
-                    >
-                      {!isCollapsed && unreadNotificationsCount}
-                    </Badge>
-                  </Container>
-                </Link>
-              </Container>
-
-              <Container className={style.menuItem}>
-                <Link as={ReactRouterLink} to={`/perfil/${user.id}`}>
-                  <Icon as={UserIcon} fill="gray.700" />
-                  {!isCollapsed && <Text>Perfil</Text>}
-                </Link>
-              </Container>
-
-              <Container className={[style.menuItem, style.soon].join(' ')}>
-                <Link href="#">
-                  <Container className={style.menuLinkWrapper} gap={14}>
-                    <Icon as={PodiumIcon} fill="gray.700" />
-                    {!isCollapsed && <Text>Ranking</Text>}
-                  </Container>
-                  {!isCollapsed && (
-                    <Container justify="end">
-                      <Badge
-                        fontSize={{ base: 10, md: 8, '2xl': 10 }}
-                        paddingBottom={{ base: '4px', md: '3px', '2xl': '4px' }}
-                      >
-                        Em breve
-                      </Badge>
-                    </Container>
-                  )}
-                </Link>
-              </Container>
-
-              <Container className={[style.menuItem, style.soon].join(' ')}>
-                <Link href="#">
-                  <Container className={style.menuLinkWrapper} gap={14}>
-                    <Icon as={ShopIcon} fill="gray.700" />
-                    {!isCollapsed && <Text>Loja</Text>}
-                  </Container>
-                  {!isCollapsed && (
-                    <Container justify="end">
-                      <Badge
-                        fontSize={{ base: 10, md: 8, '2xl': 10 }}
-                        paddingBottom={{ base: '4px', md: '3px', '2xl': '4px' }}
-                      >
-                        Em breve
-                      </Badge>
-                    </Container>
-                  )}
-                </Link>
-              </Container>
-
-              <Container className={[style.menuItem, style.soon].join(' ')}>
-                <Link href="#">
-                  <Container className={style.menuLinkWrapper} gap={14}>
-                    <Icon as={BackpackIcon} fill="gray.700" />
-                    {!isCollapsed && <Text>Inventário</Text>}
-                  </Container>
-                  {!isCollapsed && (
-                    <Container justify="end">
-                      <Badge
-                        fontSize={{ base: 10, md: 8, '2xl': 10 }}
-                        paddingBottom={{ base: '4px', md: '3px', '2xl': '4px' }}
-                      >
-                        Em breve
-                      </Badge>
-                    </Container>
-                  )}
-                </Link>
-              </Container>
+              {topMenuItems.map((item) => (
+                <SidebarMenuItem
+                  key={item}
+                  item={item}
+                  receivedInvites={receivedInvites}
+                  unreadNotifications={unreadNotifications}
+                  onClickFunction={onClickFunction}
+                />
+              ))}
             </Container>
 
-            <Container column className={style.bottomMenu}>
-              <Container className={style.menuItem}>
-                <Link as={ReactRouterLink} to="/conta">
-                  <Icon as={SettingsIcon} fill="gray.700" />
-                  {!isCollapsed && <Text>Conta</Text>}
-                </Link>
-              </Container>
-
-              <Container className={style.menuItem}>
-                <Link as="button" onClick={handleOpenModalSupport}>
-                  <Icon as={SupportIcon} fill="gray.700" />
-                  {!isCollapsed && <Text fontSize={14}>Suporte</Text>}
-                </Link>
-              </Container>
-
-              <Container className={style.menuItem}>
-                <Link as="button" onClick={handleOpenModalLogout}>
-                  <Icon as={ExitIcon} fill="gray.700" />
-                  {!isCollapsed && <Text fontSize={14}>Sair</Text>}
-                </Link>
-              </Container>
+            <Container className={style.bottomMenu} column>
+              {bottomMenuItems.map((item) => (
+                <SidebarMenuItem
+                  key={item}
+                  item={item}
+                  receivedInvites={receivedInvites}
+                  unreadNotifications={unreadNotifications}
+                  onClickFunction={onClickFunction}
+                />
+              ))}
             </Container>
           </Container>
         </Container>
 
-        <Container className={style.footer} align="end">
-          {isCollapsed ? (
-            <Container
-              justify="center"
-              gap={24}
-              style={{ paddingLeft: '16px', paddingRight: '16px' }}
-            >
-              <Icon as={ShareIcon} color="gray.700" fontSize={18} />
-            </Container>
-          ) : (
-            <Container gap={24}>
-              <Link
-                href="https://www.instagram.com/reloadclubgg/"
-                isExternal
-                fontSize={{ base: 18, md: 16, '2xl': 18 }}
-              >
-                <SiInstagram />
-              </Link>
-
-              <Link
-                href="https://twitter.com/reloadclubgg"
-                isExternal
-                fontSize={{ base: 18, md: 16, '2xl': 18 }}
-              >
-                <SiTwitter />
-              </Link>
-
-              <Link
-                href="https://discord.gg/mMMKshktfT"
-                isExternal
-                fontSize={{ base: 18, md: 16, '2xl': 18 }}
-              >
-                <SiDiscord />
-              </Link>
-
-              <Link
-                href="https://www.youtube.com/channel/UC0Yx6OapSWC0pym9ACd-D1A"
-                isExternal
-                fontSize={{ base: 18, md: 16, '2xl': 18 }}
-              >
-                <SiYoutube />
-              </Link>
-
-              <Link
-                href="https://www.facebook.com/profile.php?id=100089787770305"
-                isExternal
-                fontSize={{ base: 18, md: 16, '2xl': 18 }}
-              >
-                <SiFacebook />
-              </Link>
-            </Container>
-          )}
-        </Container>
+        <SidebarFooter />
       </Container>
 
       <NotificationList